+++ docs/changelog.rst
@@ -14,6 +14,7 @@ v1.1.0 - not yet released
 * 304: Allow users to upload profile image - http://github.com/mpowering/django-orb/issues/issue/304
 * 287: New field for attribution for resources - http://github.com/mpowering/django-orb/issues/issue/287
 * 80: Provide analytics of resource usage/downloads to content providers - http://github.com/mpowering/django-orb/issues/issue/80
+* 270: Add no resources submitted during month on Visitor analytics page - http://github.com/mpowering/django-orb/issues/issue/270
 
 v1.0.1
 ---------------------